Error handling: Remove errbacks, instead return error and result objects
This commit removes the need of individual errback functions. Every
succesful result is an instance of a subclass of the TrombiObject and
has instance variable error which is set to False. If the call results
in error a TrombiError object is returned. A TrombiError object has
instance variable error set to True.